Minor update towards 1.1.0 release.
git-svn-id: http://svn.pyopt.org/trunk@19 1cfd4a00-6b3e-4059-ac93-9806997b0ed0

+++ pyOpt/pyCOBYLA/pyCOBYLA.py
@@ -343,7 +343,7 @@ class COBYLA(Optimizer):
 		#end
 		maxfun = numpy.array([self.options['MAXFUN'][1]], numpy.int)
 		w = numpy.zeros([n*(3*n+2*m+11)+4*m+6], numpy.float)
-		iact = numpy.zeros([m+1], numpy.int)
+		iact = numpy.zeros([m+1], numpy.intc)
 		ifail = numpy.array([0], numpy.int)
 		nfvals = numpy.array([0], numpy.int)
 		iout = numpy.array([self.options['IOUT'][1]], numpy.int)
